correcting more errors

+++ ros2_dual_stepper_controller/include/ros2_dual_stepper_controller/wheel_joint.hpp
@@ -7,18 +7,18 @@
 struct WheelJoint {
   std::string name;
 
-  float commanded_velocity_rad_s = 0.0;
+  double commanded_velocity_rad_s = 0.0;
 
-  float position_rad = 0.0;               // current raw encoder reading
-  float previous_position_rad = 0.0;      // previous position in radians
+  double position_rad = 0.0;               // current raw encoder reading
+  double previous_position_rad = 0.0;      // previous position in radians
 
-  float total_position_rad = 0.0;         // unwrapped continuous position
-  float last_total_position_rad = 0.0;    // last total unwrapped position
+  double total_position_rad = 0.0;         // unwrapped continuous position
+  double last_total_position_rad = 0.0;    // last total unwrapped position
 
-  float velocity_rad_s = 0.0;             // current angular velocity in rad/s
+  double velocity_rad_s = 0.0;             // current angular velocity in rad/s
 
-  float updateTotalPositionRad();
-  void computeVelocityRadS(float dt);
+  double updateTotalPositionRad();
+  void computeVelocityRadS(double dt);
 
 };
