+++ ros2_dual_stepper_controller/src/wheel_joint.cpp
@@ -4,8 +4,8 @@
 
 
 // Update the total position and manage wrap around
-float WheelJoint::updateTotalPositionRad() {
-    float delta = position_rad - previous_position_rad;
+double WheelJoint::updateTotalPositionRad() {
+    double delta = position_rad - previous_position_rad;
 
     if (std::abs(delta) > M_PI) {
         if (delta > 0) delta -= 2 * M_PI;
@@ -20,8 +20,8 @@ float WheelJoint::updateTotalPositionRad() {
 
 
 // Compute the angular velocity in rad/s
-void WheelJoint::computeVelocityRadS(float dt) {
-    float delta = total_position_rad - last_total_position_rad;
+void WheelJoint::computeVelocityRadS(double dt) {
+    double delta = total_position_rad - last_total_position_rad;
     velocity_rad_s = delta / dt;
     last_total_position_rad = total_position_rad;
 }
